pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fuzzytrackmatch"
version = "0.1.0"
dependencies = ["pylast", "pyyaml"]

[tool.setuptools.packages.find]
where = ["src"]